requirements.txt
elasticsearch>=8.0.0
mcp-sdk>=0.1.0
